fixed some dum dum problems
